requirements.txt
# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.9.0

# LangChain ecosystem (Compatible with Python 3.11-3.12)
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-pinecone==0.2.0

# Pinecone
pinecone-client==5.0.0

# ML & Embeddings
sentence-transformers==3.0.0
numpy==1.26.4
pandas==2.2.0

# Utilities
python-dotenv==1.0.1
websockets==13.0
httpx==0.27.0
